Move transcode into jp2.py + add test
